+++ src/components/ids-date-picker/ids-date-picker.ts
@@ -326,15 +326,6 @@ class IdsDatePicker extends Base {
     this.picker = this.container?.querySelector<IdsDatePickerPopup>('ids-date-picker-popup');
     if (this.picker && this.picker.isConnected) {
       this.picker.appendToTargetParent();
-      this.picker.popupOpenEventsTarget = document.body;
-      this.picker.onOutsideClick = (e: Event) => {
-        if (this.picker) {
-          if (!e.composedPath()?.includes(this.picker)) {
-            this.triggerEvent('outsideclick.datepicker', this);
-            this.#togglePopup(false);
-          }
-        }
-      };
       this.picker.onTriggerClick = () => {
         if (this.disabled || this.readonly) return;
         this.picker?.toggleVisibility();
@@ -353,6 +344,8 @@ class IdsDatePicker extends Base {
           this.picker.popup.setAttribute(attributes.ALIGN, `bottom, ${rtlAdjustedAlignValue}`);
         }
 
+        this.picker.popup.onOutsideClick = this.onOutsideClick.bind(this);
+
         // Detect switch of X/Y values due to alignment settings,
         // and account for extra width needed to be displayed outside of IdsDatePicker fields
         this.picker.popup.onXYSwitch = onPickerPopupXYSwitch;
@@ -367,19 +360,20 @@ class IdsDatePicker extends Base {
     }
   }
 
+  onOutsideClick(e: Event) {
+    if (this.picker) {
+      if (!e.composedPath()?.includes(this.picker)) {
+        this.triggerEvent('outsideclick.datepicker', this);
+        this.#togglePopup(false);
+      }
+    }
+  }
+
   /**
    * Establish internal event handlers
    * @returns {object} The object for chaining
    */
   #attachEventHandlers(): object {
-    // Input value change triggers component value change
-    this.offEvent('change.date-picker-input');
-    this.onEvent('change.date-picker-input', this.triggerField, (e: any) => {
-      this.setAttribute(attributes.VALUE, e.detail.value);
-      this.picker?.setAttribute(attributes.VALUE, e.detail.value);
-      this.bubbleEvent(e);
-    });
-
     // Date Picker Popup's `dayselected` event causes the trigger field value to the change
     this.offEvent('dayselected.date-picker-popup');
     this.onEvent('dayselected.date-picker-popup', this.container, (e: IdsDayselectedEvent) => {