+++ src/components/ids-lookup/ids-lookup.ts
@@ -247,12 +247,8 @@ export default class IdsLookup extends Base {
         const tfValue = this.triggerField.value;
         if (tfValue !== this.state.value) {
           this.triggerField.value = this.state.value;
-          this.triggerField?.input?.dispatchEvent(new Event('change'));
         }
       }
-
-      // Send the change event
-      this.triggerEvent('change', this, { detail: { elem: this, value: this.state.value } });
     }
   }
 
@@ -380,7 +376,7 @@ export default class IdsLookup extends Base {
   #syncSelectedRows(value: string = this.value): void {
     // Deselect all rows, if given value is empty
     if (value === '') {
-      if (this.dataGrid?.selectedRows.length) this.dataGrid.deSelectAllRows();
+      if (this.dataGrid?.selectedRows?.length) this.dataGrid.deSelectAllRows();
       if (this.value !== value) this.value = value;
       return;
     }