+++ src/core/ids-attributes.ts
@@ -34,6 +34,7 @@ export const attributes = {
   AUTO_FILL: 'auto-fill',
   AUTO_FIT: 'auto-fit',
   AUTO_FLOW: 'auto-flow',
+  AUTO_FOCUS: 'auto-focus',
   AUTO_HEIGHT: 'auto-height',
   AUTO_ROWS: 'auto-rows',
   AUTO_START: 'auto-start',
@@ -572,7 +573,7 @@ export const prefix = {
   PREFIX: 'ids'
 };
 
-export const version = '1.0.0-beta.19';
+export const version = '1.0.0-beta.21';
 
 export enum IdsDirection {
   Up = 'up',